Split layout components
