+++ src/pages/stores/[id]/index.tsx
@@ -1,82 +1,25 @@
-import { HeartOutlined, HeartTwoTone } from '@ant-design/icons'
-import Image from 'next/image'
-import Link from 'next/link'
-import { useRouter } from 'next/router'
-import { createContext, ReactElement, ReactNode, useContext, useMemo } from 'react'
+import { ReactElement } from 'react'
+import { useRecoilValue } from 'recoil'
 import PageHead from 'src/components/PageHead'
-import { useStoreDetailQuery, useStoreQuery } from 'src/graphql/generated/types-and-hooks'
-
-type TStoreContext = {
-  id: string
-  name: string
-}
-
-// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-export const StoreContext = createContext<TStoreContext>(undefined!)
-
-export function StoreLayout({ children }: { children: ReactNode }) {
-  const router = useRouter()
-
-  const storeId = (router.query.id ?? '') as string
-
-  const { data, loading } = useStoreQuery({ skip: !storeId, variables: { storeId } })
-
-  const store = data?.store
-  const storeName = store?.name ?? '매장'
-
-  const storeContext = useMemo(
-    () => ({
-      id: storeId,
-      name: storeName,
-    }),
-    [storeId, storeName]
-  )
-
-  return (
-    <div>
-      <Image
-        src={store?.imageUrls?.[0] ?? '/images/default-store-cover.png'}
-        alt="store-cover"
-        width="200"
-        height="200"
-        objectFit="cover"
-      />
-      {loading || !store ? (
-        'loading...'
-      ) : (
-        <>
-          <h2>{storeName}</h2>
-          <div>{store.description}</div>
-          <div>{store.isLiked ? <HeartTwoTone twoToneColor="#ff9f74" /> : <HeartOutlined />}</div>
-        </>
-      )}
-      <div>
-        <Link href={`/stores/${storeId}/info`}>정보</Link>{' '}
-        <Link href={`/stores/${storeId}/news`}>소식</Link>{' '}
-        <Link href={`/stores/${storeId}/feed`}>피드</Link>{' '}
-        <Link href={`/stores/${storeId}/menus`}>메뉴</Link>{' '}
-      </div>
-      <StoreContext.Provider value={storeContext}>{children}</StoreContext.Provider>
-    </div>
-  )
-}
+import { useStoreDetailQuery } from 'src/graphql/generated/types-and-hooks'
+import StoreLayout from 'src/layouts/StoreLayout'
+import { store } from 'src/models/recoil'
 
 const description = ''
 
 export default function StoreInfoPage() {
-  const storeContext = useContext(StoreContext)
-  const storeId = storeContext.id
-  const storeName = storeContext.name
+  const { id: storeId, name: storeName } = useRecoilValue(store)
 
   const { data, loading, error } = useStoreDetailQuery({ skip: !storeId, variables: { storeId } })
+
   const storeDetail = data?.store
 
   return (
     <PageHead title={`${storeName} 정보 - 소복`} description={description}>
       <div>매장 정보 페이지</div>
-      {loading || !storeDetail ? (
+      {loading ? (
         'loading...'
-      ) : (
+      ) : storeDetail ? (
         <>
           <div>{storeDetail.tel}</div>
           <div>{storeDetail.address}</div>
@@ -89,6 +32,8 @@ export default function StoreInfoPage() {
           <div>{storeDetail.categories}</div>
           <div>{storeDetail.hashtags}</div>
         </>
+      ) : (
+        ''
       )}
     </PageHead>
   )
