+++ src/pages/stores/[id]/news.tsx
@@ -1,15 +1,15 @@
 import { ReactElement, useContext } from 'react'
+import { useRecoilValue } from 'recoil'
 import PageHead from 'src/components/PageHead'
 import StoreNewsCard from 'src/components/StoreNewsCard'
 import { useStoreNewsQuery } from 'src/graphql/generated/types-and-hooks'
-import { StoreContext, StoreLayout } from '.'
+import StoreLayout from 'src/layouts/StoreLayout'
+import { store } from 'src/models/recoil'
 
 const description = ''
 
 export default function StoreNewsPage() {
-  const storeContext = useContext(StoreContext)
-  const storeId = storeContext.id
-  const storeName = storeContext.name
+  const { id: storeId, name: storeName } = useRecoilValue(store)
 
   const { data, loading, error } = useStoreNewsQuery({ skip: !storeId, variables: { storeId } })
